requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning & NLP
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Vector Database
weaviate-client==4.4.0

# Environment & Configuration
python-dotenv==1.0.0

# Job Scheduling
joblib==1.3.2

# OpenAI (GPT 번역용)
openai==1.3.7

